safari-534.53-branch/LayoutTests/http/tests/websocket/tests/close-on-unload-reference-in-parent.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<link rel="stylesheet" href="../../../js-test-resources/js-test-style.css">
<script src="../../../js-test-resources/js-test-pre.js"></script>
<script src="../../../js-test-resources/js-test-post-function.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description("Test if Web Socket opened in iframe but referred in parent document is closed on unloading the frame.");

if (window.layoutTestController)
    layoutTestController.waitUntilDone();

var frameDiv;
var closedSocket;
document.childWebSocket = null;

function endTest()
{
    shouldBe("closedSocket", '"socket1"');
    shouldBe("document.childWebSocket.readyState", "2");
    isSuccessfullyParsed();
    if (window.layoutTestController)
        layoutTestController.notifyDone();
};

var ws = new WebSocket("ws://127.0.0.1:8880/websocket/tests/close-on-unload");
ws.onopen = function()
{
    debug("PASS ws on master document is ready.");
    frameDiv = document.createElement("iframe");
    frameDiv.src = "resources/close-on-unload-iframe-reference-in-parent.html";
    document.body.appendChild(frameDiv);
    debug("PASS insert a iframe, where open ws called 'socket1'");
};
ws.onmessage = function(evt)
{
    closedSocket = evt.data;
    ws.close();
};
ws.onclose = function()
{
    endTest();
};

document.iframeReady = function()
{
    shouldBeNonNull("document.childWebSocket");
    debug("PASS 'socket1' is sent to the server. unload the iframe. expect receiving 'socket1' on ws...");
    document.body.removeChild(frameDiv);
};

var successfullyParsed = true;
</script>

</body>
</html>
